requirements.txt
flask==3.1.3
pandas==3.0.2
google-api-python-client==2.194.0
python-dotenv==1.2.1
openpyxl==3.1.5
werkzeug==3.1.8
